Fix affinity in test script
You need to change the script!!!  See patch below.  All prior results
with more than 32 CPUs are bogus because the -a arguments confined the
threads to only 32 CPUs (the even-numbered ones from zero to 62).

Allow 64 CPUs.  Really needs to do the calculation based on the
stride and the number of CPUs, but a quick hack for now.

Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>